Refactor: 코드 내 재사용 컨스턴츠 및 함수 모듈화
